modbuild: ml: bug 26259: makefiles mods to support modular build
Separate checkin of the existing makefiles and cmake files that need
to be modified in order to support the new modular build system.

For the most part, the modifications to the makefiles involve
replacing some hardwired text with variables that can be overridden
on the make command line.  

And similarly, in the CMakeLists.txt files, command line options were
added so that they could be used in conditionals and overridden on
the command line.

The intent is that there should be no fucntional differences
introduced by this change (i.e. fully compatible with the previous
version).

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/staging/PostPrimary/pbbam/": change = 148718]