build: ml: bug 26194: mods for modular build system support
Checking in the modifications to existing files that need to change
to support the modular build system checkin (coming shortly).  Just
checking in the modifications to existing files separately to reduce
the noise in case there is any issue in building the existing build
(the rest of the modular build system will just be added files).

The intent is that these changes should not affect the existing
developer or nightly builds.

Changes include:
  * enable building and using of shared libs in the CMakefile.txt
    files for bam2fastx,  optionparser, and pbbam
  * pass in the output dir to cmakefiles (for modular/nightly
    builds only)
  * disable use of rpath in cmakefiles  (for modular/nightly
    builds only)
  * get rid of chef references in CMakefile.txt files, use 'modbuild'
    instead
  * slightly better sharing of code/variables betwen modbuild and 
    dev builds
  * modify buildctl.sh files for htslib, pbbam, optionparser and
    bam2fastx and pkgbuild.sh file for zlib in order to:
      - enable modular build system
      - pass output dir to cmake
      - enable building and linking to shared libraries
      - create wrapper scripts for executables to point to shared libs

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/staging/PostPrimary/pbbam/": change = 152804]