Fixed performance issues
The previous version was being pessimistic about reads outside of the
main loop, and inside the loop was accidentally not reading anything.
Now the reads inside the loop trail the tip of the file by ~100 blocks,
and is doing a multi-threaded read in order to keep up and make sure
file caching spares us the IO hit.